+++ src/test/cpp/core/SegmentImplTest.cpp
@@ -121,9 +121,9 @@ void SegmentImplTest::testMoveForward() {
     valarray<uint32_t>& data = segment->getAccessor("U").getUIntData();
 
     const Grid& grid = segment->getGrid();
-    for (size_t k = 0; k < grid.getSizeK(); k++) {
-        for (size_t l = 0; l < grid.getSizeL(); l++) {
-            for (size_t m = 0; m < grid.getSizeM(); m++) {
+    for (long k = 0; k < grid.getSizeK(); k++) {
+        for (long l = 0; l < grid.getSizeL(); l++) {
+            for (long m = 0; m < grid.getSizeM(); m++) {
                 const size_t i = grid.getIndex(k, l, m);
                 data[i] = l;
             }
@@ -133,9 +133,9 @@ void SegmentImplTest::testMoveForward() {
     segment->moveForward(200);
     CPPUNIT_ASSERT(grid.getFirstL() == 200);
 
-    for (size_t k = 0; k < grid.getSizeK(); k++) {
-        for (size_t l = grid.getFirstL(); l < grid.getFirstL() + grid.getSizeL(); l++) {
-            for (size_t m = 0; m < grid.getSizeM(); m++) {
+    for (long k = 0; k < grid.getSizeK(); k++) {
+        for (long l = grid.getFirstL(); l < grid.getFirstL() + grid.getSizeL(); l++) {
+            for (long m = 0; m < grid.getSizeM(); m++) {
                 const size_t i = grid.getIndex(k, l, m);
                 if (l < grid.getFirstL() + grid.getSizeL() - 200) {
                     CPPUNIT_ASSERT(data[i] == l);
@@ -155,9 +155,9 @@ void SegmentImplTest::testMoveForward() {
     segment->moveForward(2200);
     CPPUNIT_ASSERT(grid.getFirstL() == 2200);
 
-    for (size_t k = 0; k < grid.getSizeK(); k++) {
-        for (size_t l = grid.getFirstL(); l < grid.getFirstL() + grid.getSizeL(); l++) {
-            for (size_t m = 0; m < grid.getSizeM(); m++) {
+    for (long k = 0; k < grid.getSizeK(); k++) {
+        for (long l = grid.getFirstL(); l < grid.getFirstL() + grid.getSizeL(); l++) {
+            for (long m = 0; m < grid.getSizeM(); m++) {
                 const size_t i = grid.getIndex(k, l, m);
                 CPPUNIT_ASSERT(data[i] == 0);
             }
